confi_scalper/requirements.txt
# Асинхронная работа и HTTP
aiohttp>=3.9.0
websockets>=12.0

# Конфигурация
pyyaml>=6.0
python-dotenv>=1.0.0

# Логирование
loguru>=0.7.0

# Тесты
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Утилиты
python-dateutil>=2.8.2
